+++ service/zpages.go
@@ -23,6 +23,7 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/internal/version"
+	"go.opentelemetry.io/collector/service/featuregate"
 	"go.opentelemetry.io/collector/service/internal/zpages"
 )
 
@@ -31,6 +32,7 @@ const (
 	servicezPath   = "servicez"
 	pipelinezPath  = "pipelinez"
 	extensionzPath = "extensionz"
+	featurezPath   = "featurez"
 
 	zPipelineName  = "zpipelinename"
 	zComponentName = "zcomponentname"
@@ -42,6 +44,7 @@ func (srv *service) RegisterZPages(mux *http.ServeMux, pathPrefix string) {
 	mux.Handle(path.Join(pathPrefix, tracezPath), otelzpages.NewTracezHandler(srv.zPagesSpanProcessor))
 	mux.HandleFunc(path.Join(pathPrefix, servicezPath), srv.handleServicezRequest)
 	mux.HandleFunc(path.Join(pathPrefix, pipelinezPath), srv.handlePipelinezRequest)
+	mux.HandleFunc(path.Join(pathPrefix, featurezPath), handleFeaturezRequest)
 	mux.HandleFunc(path.Join(pathPrefix, extensionzPath), func(w http.ResponseWriter, r *http.Request) {
 		handleExtensionzRequest(srv, w, r)
 	})
@@ -58,6 +61,11 @@ func (srv *service) handleServicezRequest(w http.ResponseWriter, r *http.Request
 	zpages.WriteHTMLComponentHeader(w, zpages.ComponentHeaderData{
 		Name:              "Extensions",
 		ComponentEndpoint: extensionzPath,
+		Link:              true,
+	})
+	zpages.WriteHTMLComponentHeader(w, zpages.ComponentHeaderData{
+		Name:              "Features",
+		ComponentEndpoint: featurezPath,
 		Link:              true,
 	})
 	zpages.WriteHTMLPropertiesTable(w, zpages.PropertiesTableData{Name: "Build And Runtime", Properties: version.RuntimeVar()})
@@ -149,5 +157,25 @@ func getExtensionsSummaryTableData(host component.Host) zpages.SummaryExtensions
 	sort.Slice(data.Rows, func(i, j int) bool {
 		return data.Rows[i].FullName < data.Rows[j].FullName
 	})
+	return data
+}
+
+func handleFeaturezRequest(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/html; charset=utf-8")
+	zpages.WriteHTMLPageHeader(w, zpages.HeaderData{Title: "Feature Gates"})
+	zpages.WriteHTMLFeaturesTable(w, getFeaturesTableData())
+	zpages.WriteHTMLPageFooter(w)
+}
+
+func getFeaturesTableData() zpages.FeatureGateTableData {
+	data := zpages.FeatureGateTableData{}
+	for _, g := range featuregate.List() {
+		data.Rows = append(data.Rows, zpages.FeatureGateTableRowData{
+			ID:          g.ID,
+			Enabled:     g.Enabled,
+			Description: g.Description,
+		})
+	}
+
 	return data
 }